Add a basic DB backup solution
This is very basic, but it protects against the box going down and
corrupting memory. This backup goroutine sleeps for a time duration
between backups, overwriting a backup file on disk at each interval.

The default baked-in backup interval is 300 seconds. It should be a
config option some time soon.