Document container, show how to persist godb.json on hosting system
This modifies the way the link database is being imported to handle
errors. The container is built to allow a volume for the purpose of storing
godb.json to disk on the hosting system. If the volume isn't used, the
redirector just uses a new one each time the container is run.