Implement basic active-standby failover (#13)
* Add a basic synchronization mechanism between active/standby

This just implements the initial handshake to see who becomes active in
the pair.

* Implement simple active/standby failover

This is a crude active/standby pair implementation. It is designed
specifically to allow the redirector to be worked on or updated while
another takes traffic, both behind a load balancer. The active system
will have its webserver up, so it can be health-checked normally by the
LB. If/when it goes down, the standby will detect a loss of heartbeat
and go active with the most recent version of the link database synced
over from the (now missing) active redirector.

Co-authored-by: Bill Booth <cwbooth5@gmail.com>