Add checkmode, maps, and strings
This adds three new features and a bunch of bug fixes and improvements.

Check mode allows users to see what the redirector is doing internally
as it looks up information and replaces strings in URLs.

The maps/strings feature allows regex named capture groups to be used to
perform map lookups within the redirector. This is a relatively advanced
feature and the implementation is a bit rough, but it does work well.
This allows user input parameters to dictate the final URL the
redirector generates.

Race conditions were fixed with a repositioning of the shutown code and
some locks around database access.